Getting the infrastructure fot NetFX test runner inplace
This change will enable running the tests with NetFX.
Note that we still need to go through all tests and fix any issues there like assembly binding issues or the test is written in way that not suitable for the desktop runs

The change also include some fixes in return the src and test sources compile back with NetFx
