+++ Tools-Override/XUnitNetFXConfigFileTemplate.txt
@@ -1,0 +1,74 @@
+<?xml version="1.0"?>
+<configuration>
+   <runtime>  
+      <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1"> 
+
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.execution.dotnet" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.2.0.3300" href="{RunTimePath}xunit.execution.dotnet.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="System.IO.FileSystem" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+            <bindingRedirect oldVersion="4.0.0.0" newVersion="4.0.1.0"/>
+            <codeBase version="4.0.2.0" href="{RunTimePath}System.IO.FileSystem.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="Xunit.NetCore.Extensions" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
+            <codeBase version="999.999.999.999" href="{RunTimePath}Xunit.NetCore.Extensions.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.core" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.2.0.3300" href="{RunTimePath}xunit.core.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.assert" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.2.0.3300" href="{RunTimePath}xunit.assert.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.execution.desktop" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.2.0.3300" href="{RunTimePath}xunit.execution.desktop.dll"/>
+         </dependentAssembly>
+      
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.abstractions" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.0.0.0" href="{RunTimePath}xunit.abstractions.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="xunit.runner.utility.desktop" publicKeyToken="8d05b1bb7a6fdb6c" culture="neutral" />
+            <codeBase version="2.2.0.3444" href="{RunTimePath}xunit.runner.utility.desktop.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>
+            <assemblyIdentity name="netstandard" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
+            <codeBase version="2.0.0.0" href="{RunTimePath}netstandard.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>
+            <assemblyIdentity name="System.IO.Ports" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+            <codeBase version="4.0.0.0" href="{RunTimePath}System.IO.Ports.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>  
+            <assemblyIdentity name="System.Runtime.InteropServices.RuntimeInformation" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+            <bindingRedirect oldVersion="4.0.0.0" newVersion="4.0.2.0"/>
+            <codeBase version="4.0.2.0" href="{RunTimePath}System.Runtime.InteropServices.RuntimeInformation.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>
+            <assemblyIdentity name="System.Text.Encoding.CodePages" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+            <codeBase version="4.1.0.0" href="{RunTimePath}System.Text.Encoding.CodePages.dll"/>
+         </dependentAssembly>
+
+         <dependentAssembly>
+            <assemblyIdentity name="Microsoft.Win32.Registry" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+            <codeBase version="4.1.0.0" href="{RunTimePath}Microsoft.Win32.Registry.dll"/>
+         </dependentAssembly>
+      </assemblyBinding>
+   </runtime>
+</configuration>
