+++ src/System.Linq/tests/SelectManyTests.cs
@@ -9,7 +9,7 @@ using Xunit;
 
 namespace System.Linq.Tests
 {
-    public class SelectManyTests : EnumerableTests
+    public partial class SelectManyTests : EnumerableTests
     {
         [Fact]
         public void EmptySource()
@@ -474,45 +474,6 @@ namespace System.Linq.Tests
         {
             var iterator = counts.SelectMany(c => Enumerable.Range(1, c));
             Assert.Throws<OverflowException>(() => iterator.Count());
-        }
-
-        [Theory]
-        [InlineData(10)]
-        public void EvaluateSelectorOncePerItem(int count)
-        {
-            int[] timesCalledMap = new int[count];
-
-            IEnumerable<int> source = Enumerable.Range(0, 10);
-            IEnumerable<int> iterator = source.SelectMany(index =>
-            {
-                timesCalledMap[index]++;
-                return new[] { index };
-            });
-            
-            // Iteration
-            foreach (int index in iterator)
-            {
-                Assert.Equal(Enumerable.Repeat(1, index + 1), timesCalledMap.Take(index + 1));
-                Assert.Equal(Enumerable.Repeat(0, timesCalledMap.Length - index - 1), timesCalledMap.Skip(index + 1));
-            }
-
-            Array.Clear(timesCalledMap, 0, timesCalledMap.Length);
-
-            // ToArray
-            iterator.ToArray();
-            Assert.Equal(Enumerable.Repeat(1, timesCalledMap.Length), timesCalledMap);
-
-            Array.Clear(timesCalledMap, 0, timesCalledMap.Length);
-
-            // ToList
-            iterator.ToList();
-            Assert.Equal(Enumerable.Repeat(1, timesCalledMap.Length), timesCalledMap);
-
-            Array.Clear(timesCalledMap, 0, timesCalledMap.Length);
-
-            // ToHashSet
-            iterator.ToHashSet();
-            Assert.Equal(Enumerable.Repeat(1, timesCalledMap.Length), timesCalledMap);
         }
     }
 }
