+++ src/System.Text.Encoding.CodePages/src/Configurations.props
@@ -5,6 +5,7 @@
       netcoreapp-Unix;
       netcoreapp-Windows_NT;
       uap-Windows_NT;
+      netstandard-Windows_NT;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
